requirements.txt
faiss-cpu
numpy
pandas
sentence-transformers
langchain
ipywidgets
notebook
accelerate
tabulate